examples/bench/toy/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
docker-pycreds==0.4.0
filelock==3.9.0
fsspec==2023.4.0
gdown==5.1.0
geoopt==0.5.0
gitdb==4.0.11
GitPython==3.1.42
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
pyaml==23.12.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.12.0
sentry-sdk==1.40.6
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sympy==1.12
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.5
